requirements.txt
streamlit
auth0-python